Add fallback option when image file is not in local directory
In case the image is not found in IMAGE_PATH, it will retrieve the image binary from IMAGE_URL without need for local storage.

Motivation: This code is published in other sites (e.g. https://docs.astropy.org/en/stable/generated/examples/io/split-jpeg-to-fits.html) without a link to the image file. By adding this fallback option, it increases the robustness of the code.